// src/app/model/tblField.ts
export class TblField {
  id?: number;
  provinceid?: number;
  districtId?: number;
  wardId?: number;
  createby?: string;
  modifiedby?: string;
  createdDate?: Date;
  modifiedDate?: Date;
  fieldName?: string;
  provinceName?: string;
  districtName?: string;
  wardName?: string;
  phoneNumberField?: string;
  day?: Date;
  timeStart?: Date;
  timeEnd?: Date;
  fieldType?: number;
  address?: string;
  image?: string;
  supplierId?: number;
  supplierName?: string;
  name: string;
  status?: string;
}