// src/app/model/tblShiftField.ts
export class TblShiftField{
    id?: number;
    fieldId?: number;
    shiftFieldName?: number;
    createdBy?: string;
    createdDate?: Date;
    modifiedDate?: Date;
    modifiedBy?: string;
    timeStart?: string;
    timeEnd?: string;
    amountWeekday?: string;
    amountWeekend?: string;
    dayOfWeek?: string;
    day?: Date;
    statusField?: string;
    fieldType?: string;
}