// src/app/model/tblSmallField.ts
export class TblSmallField{
    id?: number;
    fieldId?: number;
    provinceId?: number;
    districtId?: number;
    wardId?: number;
    fieldTypeId?: number;
    fieldMoney?: string;
    status?: string;
    smallFieldName?: string;
    fieldType?: string;
    createdBy?: string;
    createdDate?: Date;
    modifiedDate?: Date;
    modifiedBy?: string;
}